+++ pkg/analyzer/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Deprecated `LibraryElement.exportedLibraries`.
 * Deprecated `LibraryElement.importedLibraries`.
 * Deprecated `LibraryElement.isBrowserApplication`.
+* Deprecated `LibraryElement.libraryExports`,
+  use `CompilationUnitElement.libraryExports` instead.
 
 ## 6.9.0
 * `NormalFormalParameter` now implements `AnnotatedNode`.